pyproject.toml
[tool.poetry]
name = "finance-analyzer"
version = "0.1.0"
description = "Financial transactions analyzer"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.0"
requests = "^2.27.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
line_length = 119
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"